requirements.txt
pandas==1.4.4
Pillow==9.4.0
streamlit==1.19.0
yfinance==0.2.12
